Add throttling policy template for AI APIs

+++ components/apimgt/org.wso2.carbon.apimgt.gateway/src/main/java/org/wso2/carbon/apimgt/gateway/throttling/publisher/DataProcessAndPublishingAgent.java
@@ -71,9 +71,9 @@ public class DataProcessAndPublishingAgent implements Runnable {
     String apiName;
     String appId;
     String ipAddress;
-    Integer totalTokens;
-    Integer promptTokens;
-    Integer completionTokens;
+    Long totalTokens;
+    Long promptTokens;
+    Long completionTokens;
     Map<String, String> headersMap;
     Map<String, Object> customPropertyMap;
     private AuthenticationContext authenticationContext;
